Add error boundary
Update error boundary component
