+++ client/src/ErrorBoundary.jsx
@@ -1,0 +1,42 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+class ErrorBoundary extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = { hasError: false, error: '', errorInfo: '' };
+  }
+
+  static getDerivedStateFromError(error) {
+    // Update state so the next render will show the fallback UI.
+    return { hasError: true, error };
+  }
+
+  componentDidCatch(error, errorInfo) {
+    // TODO: Log the error to an error reporting service
+    // logErrorToMyService(error, errorInfo);
+    this.setState({ hasError: true, error, errorInfo });
+  }
+
+  render() {
+    const { hasError } = this.state;
+    const { children } = this.props;
+    if (hasError) {
+      // Error path
+      return (
+        <div>
+          <h2>Something went wrong.</h2>
+        </div>
+      );
+    }
+
+    // Normally, just render children
+    return children;
+  }
+}
+
+ErrorBoundary.propTypes = {
+  children: PropTypes.node.isRequired,
+};
+
+export default ErrorBoundary;
